Simplify test harness
