+++ test/bigtest/helpers/setup-application.js
@@ -1,0 +1,11 @@
+import setupStripesCore from '@folio/stripes-core/test/bigtest/helpers/setup-application';
+import mirageOptions from '../network';
+
+export default function setupApplication({
+  scenarios
+} = {}) {
+  setupStripesCore({
+    mirageOptions,
+    scenarios
+  });
+}
